0.0.58 Pre-alpha - Now all actions in the API are differentiated by the rights of user groups. Previously, it was possible to perform absolutely all actions regardless of the availability of rights; - Added a new right for administration: manage CMS feeds; - Minor fixes;

+++ api/media/delete.handler.php
@@ -14,30 +14,40 @@ if (!defined('IS_NOT_HACKED')) {
 }
 
 if ($system_core->client->is_logged(2)) {
-  $file_fullname = $_DELETE['media_file_fullname'];
+  $client_user = $system_core->client->get_user(2);
+  $client_user->init_data(['metadata_json']);
+  $client_user_group = $client_user->get_group();
+  $client_user_group->init_data(['permissions']);
 
-  if (isset($_DELETE['media_file_fullname'])) {
-    $media_dir_path = sprintf('%s/uploads/media', CMS_ROOT_DIRECTORY);
-    $media_file_dir_path = sprintf('%s/%s', $media_dir_path, $file_fullname);
+  if ($client_user_group->permission_check($client_user_group::PERMISSION_EDITOR_MEDIA_FILES_MANAGEMENT)) {
+    $file_fullname = $_DELETE['media_file_fullname'];
 
-    if (file_exists($media_file_dir_path)) {
-      unlink($media_file_dir_path);
+    if (isset($_DELETE['media_file_fullname'])) {
+      $media_dir_path = sprintf('%s/uploads/media', CMS_ROOT_DIRECTORY);
+      $media_file_dir_path = sprintf('%s/%s', $media_dir_path, $file_fullname);
 
-      if (!file_exists($media_file_dir_path)) {
-        $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_FILE_SUCCESS') : $handler_message;
-        $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+      if (file_exists($media_file_dir_path)) {
+        unlink($media_file_dir_path);
+
+        if (!file_exists($media_file_dir_path)) {
+          $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_FILE_SUCCESS') : $handler_message;
+          $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+        } else {
+          $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+          $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+        }
       } else {
-        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_FILE_ERROR_NOT_FOUND')) : $handler_message;
         $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
       }
     } else {
+      http_response_code(400);
       $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_FILE_ERROR_NOT_FOUND')) : $handler_message;
       $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
     }
   } else {
-    http_response_code(400);
-    $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_FILE_ERROR_NOT_FOUND')) : $handler_message;
-    $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+    $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_DONT_HAVE_PERMISSIONS'));
+    $handler_status_code = 0;
   }
 } else {
   http_response_code(401);
