+++ api/pageStatic/delete.handler.php
@@ -16,27 +16,37 @@ if (!defined('IS_NOT_HACKED')) {
 use \core\PHPLibrary\PageStatic as PageStatic;
 
 if ($system_core->client->is_logged(2)) {
-  if (isset($_DELETE['page_static_id'])) {
-    $page_static_id = (is_numeric($_DELETE['page_static_id'])) ? (int)$_DELETE['page_static_id'] : 0;
+  $client_user = $system_core->client->get_user(2);
+  $client_user->init_data(['metadata_json']);
+  $client_user_group = $client_user->get_group();
+  $client_user_group->init_data(['permissions']);
 
-    if (PageStatic::exists_by_id($system_core, $page_static_id)) {
-      $page_static = new PageStatic($system_core, $page_static_id);
-      $page_static_is_deleted = $page_static->delete();
+  if ($client_user_group->permission_check($client_user_group::PERMISSION_EDITOR_PAGES_STATIC_EDIT)) {
+    if (isset($_DELETE['page_static_id'])) {
+      $page_static_id = (is_numeric($_DELETE['page_static_id'])) ? (int)$_DELETE['page_static_id'] : 0;
 
-      if ($page_static_is_deleted) {
-        $handler_message = $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS');
-        $handler_status_code = 1;
+      if (PageStatic::exists_by_id($system_core, $page_static_id)) {
+        $page_static = new PageStatic($system_core, $page_static_id);
+        $page_static_is_deleted = $page_static->delete();
+
+        if ($page_static_is_deleted) {
+          $handler_message = $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS');
+          $handler_status_code = 1;
+        } else {
+          $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN'));
+          $handler_status_code = 0;
+        }
       } else {
-        $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN'));
+        $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_STATIC_PAGE_ERROR_NOT_FOUND'));
         $handler_status_code = 0;
       }
-    } else {
-      $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_STATIC_PAGE_ERROR_NOT_FOUND'));
-      $handler_status_code = 0;
-    }
 
-    $handler_output_data['modalClose'] = true;
-    $handler_output_data['reload'] = true;
+      $handler_output_data['modalClose'] = true;
+      $handler_output_data['reload'] = true;
+    }
+  } else {
+    $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_DONT_HAVE_PERMISSIONS'));
+    $handler_status_code = 0;
   }
 } else {
   http_response_code(401);
