+++ api/template/post.handler.php
@@ -14,52 +14,62 @@ if (!defined('IS_NOT_HACKED')) {
 }
 
 if ($system_core->client->is_logged(2)) {
-  $template_name = $_POST['template_name'];
-  $template_url = 'https://repository.cms-girvas.ru/templates/' . $_POST['template_name'];
+  $client_user = $system_core->client->get_user(2);
+  $client_user->init_data(['metadata_json']);
+  $client_user_group = $client_user->get_group();
+  $client_user_group->init_data(['permissions']);
 
-  $ch = curl_init($template_url);
-  curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-  $curl_exucute_result = json_decode(curl_exec($ch), true);
-  curl_close($ch);
+  if ($client_user_group->permission_check($client_user_group::PERMISSION_ADMIN_TEMPLATES_MANAGEMENT)) {
+    $template_name = $_POST['template_name'];
+    $template_url = 'https://repository.cms-girvas.ru/templates/' . $_POST['template_name'];
 
-  if (!empty($curl_exucute_result['outputData'])) {
-    $template_dir_path = sprintf('%s/templates/%s', CMS_ROOT_DIRECTORY, $_POST['template_name']);
-    $template_archive_path = sprintf('%s/templates/%s.zip', CMS_ROOT_DIRECTORY, $_POST['template_name']);
+    $ch = curl_init($template_url);
+    curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
+    $curl_exucute_result = json_decode(curl_exec($ch), true);
+    curl_close($ch);
 
-    $ch_archive = curl_init();
-    curl_setopt($ch_archive, CURLOPT_URL, $curl_exucute_result['outputData']['download_url']);
-    curl_setopt($ch_archive, CURLOPT_RETURNTRANSFER, 1);
-    $curl_archive_exucute_result = curl_exec($ch_archive);
-    curl_close($ch_archive);
-    
-    $file = fopen($template_archive_path, "w+");
-    fputs($file, $curl_archive_exucute_result);
-    fclose($file);
+    if (!empty($curl_exucute_result['outputData'])) {
+      $template_dir_path = sprintf('%s/templates/%s', CMS_ROOT_DIRECTORY, $_POST['template_name']);
+      $template_archive_path = sprintf('%s/templates/%s.zip', CMS_ROOT_DIRECTORY, $_POST['template_name']);
 
-    if (file_exists($template_archive_path)) {
-      $zip = new ZipArchive();
+      $ch_archive = curl_init();
+      curl_setopt($ch_archive, CURLOPT_URL, $curl_exucute_result['outputData']['download_url']);
+      curl_setopt($ch_archive, CURLOPT_RETURNTRANSFER, 1);
+      $curl_archive_exucute_result = curl_exec($ch_archive);
+      curl_close($ch_archive);
+      
+      $file = fopen($template_archive_path, "w+");
+      fputs($file, $curl_archive_exucute_result);
+      fclose($file);
 
-      if ($zip->open($template_archive_path) === true) {
-        mkdir($template_dir_path);
+      if (file_exists($template_archive_path)) {
+        $zip = new ZipArchive();
 
-        $zip->extractTo($template_dir_path);
-        $zip->close();
+        if ($zip->open($template_archive_path) === true) {
+          mkdir($template_dir_path);
 
-        unlink($template_archive_path);
+          $zip->extractTo($template_dir_path);
+          $zip->close();
 
-        $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_TEMPLATE_UPLOADED') : $handler_message;
-        $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+          unlink($template_archive_path);
+
+          $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_TEMPLATE_UPLOADED') : $handler_message;
+          $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+        } else {
+          $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+          $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+        }
       } else {
-        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNZIPPING_NOT_POSSIBLE')) : $handler_message;
         $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
       }
     } else {
-      $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNZIPPING_NOT_POSSIBLE')) : $handler_message;
+      $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_TEMPLATE_ERROR_REPOSITORY_DATA_NOT_GETTED')) : $handler_message;
       $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
     }
   } else {
-    $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_TEMPLATE_ERROR_REPOSITORY_DATA_NOT_GETTED')) : $handler_message;
-    $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+    $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_DONT_HAVE_PERMISSIONS'));
+    $handler_status_code = 0;
   }
 } else {
   $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_AUTHORIZATION')) : $handler_message;
