+++ api/user/delete.handler.php
@@ -16,27 +16,37 @@ if (!defined('IS_NOT_HACKED')) {
 use \core\PHPLibrary\User as User;
 
 if ($system_core->client->is_logged(2)) {
-  if (isset($_DELETE['user_id'])) {
-    $user_id = (is_numeric($_DELETE['user_id'])) ? (int)$_DELETE['user_id'] : 0;
+  $client_user = $system_core->client->get_user(2);
+  $client_user->init_data(['metadata_json']);
+  $client_user_group = $client_user->get_group();
+  $client_user_group->init_data(['permissions']);
 
-    if (User::exists_by_id($system_core, $user_id)) {
-      $user = new User($system_core, $user_id);
+  if ($client_user_group->permission_check($client_user_group::PERMISSION_ADMIN_USERS_MANAGEMENT)) {
+    if (isset($_DELETE['user_id'])) {
+      $user_id = (is_numeric($_DELETE['user_id'])) ? (int)$_DELETE['user_id'] : 0;
 
-      $user_is_deleted = $user->delete();
-      if ($user_is_deleted) {
-        $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
-        $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+      if (User::exists_by_id($system_core, $user_id)) {
+        $user = new User($system_core, $user_id);
+
+        $user_is_deleted = $user->delete();
+        if ($user_is_deleted) {
+          $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
+          $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+        } else {
+          $handler_message = (!isset($handler_message)) ? (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+          $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+        }
       } else {
-        $handler_message = (!isset($handler_message)) ? (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_USER_ERROR_NOT_FOUND')) : $handler_message;
         $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
       }
     } else {
-      $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_USER_ERROR_NOT_FOUND')) : $handler_message;
+      $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_INVALID_INPUT_DATA_SET')) : $handler_message;
       $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
     }
   } else {
-    $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_INVALID_INPUT_DATA_SET')) : $handler_message;
-    $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+    $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_DONT_HAVE_PERMISSIONS'));
+    $handler_status_code = 0;
   }
 } else {
   $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_AUTHORIZATION')) : $handler_message;
