+++ api/webChannel/delete.handler.php
@@ -16,19 +16,29 @@ if (!defined('IS_NOT_HACKED')) {
 use \core\PHPLibrary\WebChannel as WebChannel;
 
 if ($system_core->client->is_logged(2)) {
-  $web_channel_id = (isset($_DELETE['web_channel_id'])) ? $_DELETE['web_channel_id'] : 0;
-  $web_channel_id = (is_numeric($web_channel_id)) ? (int)$web_channel_id : 0;
+  $client_user = $system_core->client->get_user(2);
+  $client_user->init_data(['metadata_json']);
+  $client_user_group = $client_user->get_group();
+  $client_user_group->init_data(['permissions']);
 
-  if ($web_channel_id != 0) {
-    if (WebChannel::exists_by_id($system_core, $web_channel_id)) {
-      $web_channel = new WebChannel($system_core, $web_channel_id);
+  if ($client_user_group->permission_check($client_user_group::PERMISSION_ADMIN_FEEDS_MANAGEMENT)) {
+    $web_channel_id = (isset($_DELETE['web_channel_id'])) ? $_DELETE['web_channel_id'] : 0;
+    $web_channel_id = (is_numeric($web_channel_id)) ? (int)$web_channel_id : 0;
 
-      $web_channel_is_deleted = $web_channel->delete();
-      if ($web_channel_is_deleted) {
-        $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
-        $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+    if ($web_channel_id != 0) {
+      if (WebChannel::exists_by_id($system_core, $web_channel_id)) {
+        $web_channel = new WebChannel($system_core, $web_channel_id);
+
+        $web_channel_is_deleted = $web_channel->delete();
+        if ($web_channel_is_deleted) {
+          $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
+          $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+        } else {
+          $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+          $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+        }
       } else {
-        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+        $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_FEED_ERROR_NOT_FOUND')) : $handler_message;
         $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
       }
     } else {
@@ -36,8 +46,8 @@ if ($system_core->client->is_logged(2)) {
       $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
     }
   } else {
-    $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_FEED_ERROR_NOT_FOUND')) : $handler_message;
-    $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+    $handler_message = sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_DONT_HAVE_PERMISSIONS'));
+    $handler_status_code = 0;
   }
 } else {
   $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_AUTHORIZATION')) : $handler_message;
